backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
sqlalchemy>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3
bcrypt>=4.0
pdfplumber>=0.10
httpx>=0.27
python-multipart>=0.0.9
pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=5.0
fpdf2>=2.7
